chore: fix yarn start
